[ELF][Morello] Start adding tlsdesc support for Morello
The general idea for tlsdesc is that the two got entries are used for a function
pointer and its argument. The dynamic linker sets both.

All that the static linker has to do is relocate the code to point to the got
entries and create a dynamic relocation (R_MORELLO_TLSDESC).

Relaxation from Global Dynamic (GD) to Local Executable (LE) is implemented.
Relaxation from GD to Initial Executable (IE) is not implemented at the moment.

Change-Id: I1e722f7d065bfb72e151af6bda5db4fa997770ed