[Morello] Add architecture plugin to handle C64 addresses
Morello addresses use 1 bit to discriminate between A64 and C64 mode. We
need to ignore that bit whenever we do any actual operations with
memory, e.g. setting a software breakpoint, otherwise we might write at
the wrong (off-by-one) address.

This is already handled in some places in lldb via ABI::FixCodeAddress,
but that's not enough. In particular, for setting software breakpoints
correctly we also need an ArchitecturePlugin overridding
GetCallableLoadAddress and GetOpcodeLoadAddress to set or clear the
discriminating bit.

This is mostly identical to what is done for Thumb in 32-bit arm.

Change-Id: Ifed536d4d4ac2142c1801a9a97274da959fd8f91