Make __builtin_cheri_offset_increment return qualified void*
This fixes a miscompilation bug where we would generate capability loads
and stores for an under-aligned result of offset_increment(). This is
caused by clang (correctly) assuming that a builtin returning T* will
return an appropriately aligned T*, but this is not the case for e.g.
__builtin_cheri_offset_increment((long*), 1).

(cherry picked from commit 522217ab6f3f039782ad93ecc77136d9b3ca2c65)