Stop using config->hasDynSymTab to determine fragment format
The static linker was using among other factors, config->hasDynSymTab to
determine the data format of a fragment, which will cause different data formats
depending on how the executable is linked.

static executable (config->hasDynSymTab == false) : static fragment format
dynamic executable (config->hasDynSymTab == true) : dynamic fragment format

This is especially problematic when used with R_MORELLO_IRELATIVE which,
although a dynamic relocation, can be present in both static and dynamic
executables. But the ABI requires that regardless of the executable type,
the fragment pointed by R_MORELLO_IRELATIVE should be in dynamic format.

The static linker should be able to figure out the data format only by the type
of the relocation, so stop using config->hasDynSymTab for that.

Change-Id: I32331609bc797d4f13e51eddeee1f30c65ad3c7e