Handle TLS GD to LE relaxation without relying on --morello-c64-plt
The linker was relying on presence/absence of --morello-c64-plt to relax the
C64/A64 TLS GD sequence to LE. However even when --morello-c64-plt is set,
there can be input objects which contain A64 TLS relocations, which the C64
TLS sequence did not recognize and vice-versa.

So do the GD to LE relaxation without relying on --morello-c64-plt.

However, the behavior of R_AARCH64_TLSDESC_ADD_LO12 however should change
depending on if the sequence is A64 or C64. So use the R_AARCH64_TLSDESC_CALL
and R_MORELLO_TLSDESC_CALL relocations which follow R_AARCH64_TLSDESC_ADD_LO12
to determine the correct behavior.

Change-Id: I1afe6ab3b10aadc49b8f968d7c1e093d99cc2cc1