[Morello] Disable expr command in lldb for AArch64
The 'expr' command is currently not supported because we do not have
ptrace support for restoring the capability registers after code
injection. Without that, running 'expr' just clears the tags and/or
corrupts the top bits of the capability registers, thus interfering with
the whole debug session in unexpected ways. It is better to disable it
for now.

Note that this also disables the 'print' command, which is really just
an alias for 'expr'.