[Morello][lldb] Fix TestGdbRemoteSigInfo.py
In Python 3, struct.unpack() is expecting bytes, but we were passing
a str object to it, causing it to error out. The str contains integers
dumped as raw, 4-byte values, so I don't think there is any encoding
that would correctly convert them from str to bytes. Instead, we round
trip through int: using ord() to get the integer value of each byte,
then converting it back to a byte via int.to_bytes(). At this point we
could pass the bytes that we obtain into struct.unpack(), but I think
int.from_bytes() is clearer, and it saves us the trouble of building the
correct format for parsing a big or little endian int.