[NFC] Rename and move concentrateReqdAlignment to be Morello-specific
The function hard-codes parameters of the capability encoding format
(and that it's using CHERI concentrate) so is architecture-specific and
does not belong under a generic CHERI name. Were this used solely in the
backend it could remain under the AArch64 target directory, but it's
used by LLD and thus needs to be somewhere like Support still.