Tightly-bound dynamic stack allocations
Currently the underlying padded (for stack alignment) allocation is
bounded but if the requested allocation size is small we end up with
bounds of the whole allocation rather than precise tight bounds.

This is correctly implemented by the generic CheriBoundAllocas pass but
not by the AArch64Sandbox pass. This target-specific pass should be
replaced with the generic pass currently shared between CHERI-MIPS and
CHERI-RISC-V to ensure bug fixes and improvements don't get missed
between architectures, but until then this fixes the AArch64 bug.

This was found by CheriBSD's cheribsdtest-purecap (and its variants)
where 1-byte, 2-byte, 4-byte and 8-byte dynamic stack allocations were
all seen to have a length of 16.