[Morello][lldb] Skip a few tests
Skip some tests that fail when running from the Android build tree.

I'm not aware of a more fine-grained way to skip these than to skip on
all Linux platforms (I've tried skipping only when the target is
Android, but it seems we don't have that information when running the
check target).

Note that these tests pass when building outside the Android tree (which
is why we skip rather than xfail).

Details on each test can be found in this Gitlab issue:
https://git.morello-project.org/morello/llvm-project/-/issues/8