[Morello] Improve global addressing
1. DSO-internal function addresses are generated from PCC
and sealed. This avoids a load instruction and should
slightly reduce the number of relocations required.

2. Capability stubs for internal objects are merged into
a table to allow relative addressing. This allows sharing
the adrp instruction between multiple uses of the capability
table. The downside to this is that it can increase register
pressure.

3. We perform a post-legalization combine to push constant
pointer adds into adrp (this can remove add instructions).
This usually happens pre-legalization, but re-using that
machinery would require either changes to the SelectionDAG
builder or doing the load from the capability table in IR
(which would affect some other SelectionDAG transforms).