[Morello][lldb] Don't hardcode FP/CFP. NFCI.
There are many places where the frame pointer is hardcoded to x29 for
AArch64 (and respectively to c29 in a capability context). This includes
both internal variable/enum names and user-visible strings (e.g. "x29"
is considered an alternative name for "fp" rather than the other way
around).

This patch removes fp and cfp from all the register number enums and
from register infos and explicitly defines and uses x29/c29 instead in
the code. We preserve the behaviour that the name of the frame pointer
is "fp"/"cfp" rather than "xN" or "cN".

Note that since we no longer know the frame pointer statically, we need
to update the register infos with frame pointer information dynamically.
For this purpose, a helper function is added that will mark one of the
registers as the frame pointer or capability frame pointer. We call this
in every file that makes use of the register infos directly, usually in
some constructor or initialization function. One additional change is
necessary in GDBRemoteRegisterContext.cpp - this file was including
RegisterInfos_arm64.h inside a function, which made it impossible
to define the helper. We therefore had to move things one scope up.

This isn't a nice way to refactor this code, but I think it's the best
way to minimize merge pains in the future.