Fix GNU machine mode attribute support for CHERI
We need to forbid the pointer mode for integers in pure-capability C,
only permitting it on actual pointer types (which makes it rather
redundant, but we can support it so might as well appease those who
still think this attribute is a good idea) and preserving the fact that
it's a pointer rather than creating a 64-bit/128-bit plain integer. We
also need to do the same for unwind_word (which is a particularly bad
concept for split register files and for hard float ABIs). For
completeness, add a capability mode to complement the pointer mode.

Conversely, forbid integer modes for capabilities, and thus pointers in
pure-capability C, in order to avoid loss of provenance; if a plain
integer is desired, the code should be modified to use one, not a
pointer type.

Whilst here, make the implementation of getRegisterWidth clearer.
Currently PointerWidth is left as the address width for pure-capability
code which is confusing and should one day be fixed, likely by
introducing a separate PointerRange/AddressWidth. Outside of X86 (which
overrides this anyway), this is used solely for the word mode, and that
is likely intended to be a plain integer, not a capability, although
given that X86 makes it an AVX type perhaps we should make it a
capability if the intent is to use it for memcpy. This at least
preserves the current behaviour and can be revisited later.