[Morello][lldb] Fix TestBSDArchives.py
While building the binaries for the test, llvm-ar throws a warning about
building in the (default) deterministic mode, where the timestamps/uids
are set to 0. We can get rid of the warning by passing -rU in the
ARFLAGS for this test.

The real issue with the warning is that it contains bogus characters
which can't be decoded as utf-8. Decoding as latin-1 works, but that
change would affect all the tests on all the platforms in ways that
I can't predict right now, so it's probably best to settle for this
workaround.