[Morello][lldb] Add end-to-end tests with Android
Add support for end-to-end lldb tests on an FVP model running Android.

The tests live in lldb/morello-android-tests and can be run independently
of the llvm build system. Each test has its own subdirectory containing:
  * source files,
  * an Android.bp describing how to build the target for the test, and
  * a ${TESTNAME}.test file containing lldb commands and FileCheck
  expectations.

Note that we do not use RUN lines to drive the setup of the tests.
Instead, all the tests are expected to have the same setup, consisting
of the following steps:
  1. Compiling the target executable
  2. Uploading it to the model (with adb)
  3. Starting lldb-server on the model
  4. Starting lldb on the host, in batch mode, and checking its output

For more details check out the README.

This patch adds a single test to serve as an example.