[Morello] Add a pass to change the way we are creating addresses for aliases in purecap.
The motivation is that the global merge pass will replace
globals with aliases to a merged global.

The merged global address is produced by loading a capability
from a pc-relative relocation (same as before).

Safe loads/stores can use the merged global directly. Unsafe
uses still need to produce the address of the global by doing
a load from the capability table.