Fix getMorelloRequiredAlignment implementation's constants
This implementation was using the CHERI-MIPS and 64-bit CHERI-RISC-V
encoding format's constants, not the Morello one, where the latter
devotes an additional 4 bits to the bounds encoding (at the expense of
the otype range) to give 2 additional bits of precision. Moreover it had
an off-by-one error in the calculation of E due to not accounting for
having an extra leading 0 in the spec's 65-bit l. These combined to mean
we were over-aligning by up to a factor of 8, as discovered by tests
failing in CheriBSD's cheribsdtest (specifically cheribsdtest-purecap
and its variants).