[Morello] Fix the newlib libcxx tests after the merge.
newlocale returns null for some versions of newlib, so
update the configure check to test for this when testing
for locale features. Also add a separate feature (newlocale)
so we can skip two tests.

This also XFAIL any tests that use the stderr output. The
model currently only outputs to stdout.