Fix subobject bounds subscripting to not treat pointer members like arrays
Otherwise, if we're performing arithmetic on a pointer via &...[] that's
inside a nested container, we will end up looking at the container's
size and using that for the bounds of the pointer loaded from the
container, which makes no sense. Similarly, subscripting itself
shouldn't look at any of the attributes on the field when it's a
pointer.