[Morello] Fix offsets for capability array elements
When dumping arrays, we compute the offsets of the elements as multiples
of the element size. For capabilities, this does not work, because the
size is defined as 17 bytes, to include the tag. The tag however is not
part of the addresable storage, and should definitely not be included
when doing address calculations.

This patch provides a very hacky fix which just adjusts the offsets for
capabilities. This only fixes the problem for arrays. We may or may not
need similar patches for other constructs. On a cursory reading, structs
seem to get their offsets directly from clang, so they shouldn't suffer
from this issue. For transparent pointers to aggregate types, this
commit should be good enough, but there may be some corner cases that
are still broken.

A better fix would probably be to use 16 as the size of capabilities,
and explicitly add 1 to it only when we're about to allocate a buffer
that also needs to store the tag. I think this would make things
easier to reason about, but that's a large and risky refactoring that
I don't think I'll have time to do before the release (I'd need to audit
all the uses of CompilerType::GetBitSize and CompilerType::GetByteSize).
Definitely worth revisiting later though, at least in order to add more
tests, if nothing else.