[Cheri] Allow encoding of landing pads as an offset to a capability in .data.rel.ro
This means .gcc_except_table can be read-only again.

The new encoding uses 0xd as the capability marker to
distinguish from the old encoding. Both old and new
encodings are currently supported to allow various
software that needs to read the landing pad value to
be updated, and we're currently defaulting to the old
encoding.