[Morello][lldb] Fix register kinds for PCC
The PCC register was defined using a DEFINE_CAP_PCC macro, which
differed from the DEFINE_CAP macro used for most other capability
registers only in terms of the register kinds. In the case of
DEFINE_CAP_PCC, these were defined via LLDB_KIND, which set everything
to LLDB_INVALID_REGNUM (completely ignoring the generic_kind parameter).

This patch fixes the issue by using DEFINE_CAP, which correctly sets up
the expected register kinds for PCC.

Fixing the generic register kind for PCC exposes another issue in
ABISysV_arm64::GetFrameState. The logic there was that if the register
context does not have a PCC, then we don't have capabilities and the
frame state is eFrameStateSimple. If the register context does have
a PCC but we fail to read it, that would be an error and we should
return false. The problem is that when reading a plain AArch64 core
file, we will not be able to read PCC from it - however the register
context would still contain a PCC register in that case, because we
always add the capability registers to non-dynamic AArch64 register
contexts. Therefore, instead of erroring out, we make the same
assumption as before - that the frame state is simple and we're not
supposed to use capabilities.