[Morello][lldb] Set FP to c17 for descriptor ABI
Set the frame pointer to x17 and c17 for the descriptor ABI. At the
moment, we don't have support for identifying that an application has
been compiled with the descriptor ABI, so instead we do something a bit
hacky: add a new target feature. This can be enabled on the command line
with:

settings set target.aarch64-morello-desc-abi true

It needs to be called *before* launching the process, otherwise it
won't take effect. It also doesn't affect certain parts of the Python
API, as used in TestCapabilityOperation.py, and probably in other
scenarios too. We can fix these when Android supports building
descriptor ABI apps, so we can actually test something.

Needs more testing.