[Morello][lldb] Fix end-to-end tests
The tests are currently broken because lldb needs to be able to find
the source files in order to set breakpoints. The debug information uses
paths relative to the Android source tree. But since we have made it
possible to run the tests independently of an Android source tree, and
we no longer 'cd' to it in each test, lldb gets understandably lost.

This patch fixes the issue by adding an ORIGINAL_SOURCE_PREFIX variable,
which is used by the test runner to add a source mapping from the
original source locations (i.e. those relative to the Android source
tree) to the current source locations (i.e. relative to the test runner
itself).