[lld][ELF][Morello] Increase bounds of dynamically relocated executable capabilities
The executable capabilities created by dynamic relocations used the
symbol value for the base, and the symbol size as the capability limit.
This results in the executable capabilities being too narrow that they
could not call other functions.

Fix this by encoding dynamically relocated executable capabilities in a way
similar to how the cap relocs for executable capabilities are encoded in the
static linking case; i.e. set the bounds to cover the .text + .ro + .rel.ro regions.