[lld][ELF][Morello] Align fragment bases of exec symbols pointed by [I]RELATIVE relocs
The executable capabilities created by [I]RELATIVE relocations used the
symbol value for the base, and the symbol size as the capability limit.
This results in the executable capabilities being too narrow that they
could not call other functions.

Fix this by encoding fragments for these relocations in a way similar to
how the cap relocs for executable capabilities are encoded in the static
linking case; i.e. set the bounds to cover the .text + .ro + .rel.ro regions.