[Morello] Improve inttoptr codegen
When producing the null capability use a copy from czr. This
allows users which allow a czr operand to use czr directly.

When producing a constant materialize the integer constant
(which clears the capability metadata bits) and use the
super-register.

If we're not doing this on a constant use gcvalue, since
that also clears the metadata bits (and againt we don't need
to create a null capability).