[Morello] Fix remaining llvm failing regression tests.
This accounts for changes in the order of symbols printed,
some extra copies emitted at -O0 (these go away at higher
optimization levels), a change in the symbol used for
unwinding capabilities and minor changes in the output of
CHERI-Generic tests.

morello-instcombine-icmp.ll was removed because it doesn't
trigger the codepath that it used to (so isn't testing
anything).