[Morello][lldb] Fix jit-loader tests
The tests were failing because lldb wasn't hitting any breakpoints.
The breakpoints weren't added because lldb wasn't managing to find
the section for __jit_debug_register_code. This turns out to be
because the target returns 2 symbols, but the code only looks at the
first one.

This patch makes sure we look through all the symbols until we find
one whose address we can get.

I haven't really taken the time to investigate why this started
failing just now, it might be because of how lldb gets its symbols
or it might be an issue somewhere else in the toolchain.