[Morello][lldb] Skip SVE core tests
There are at least 2 issues with these tests:
1. We hit an assert when trying to read v0 because the register info
from the core file expects the SVE registers to be right after the
usual AArch64 register sets. Unfortunately, we put the capability
registers there, and the SVE registers after, so this is not trivial
to fix.
2. Even if we get the v0 right, we still get some issues related to the
precision of the floating point registers (i.e. we read values that are
almost but not quite the expected ones). This looks very similar to the
FreeBSD bug that was already reported for one of these tests.

This patch introduces a new decorator, skipIfMorello, to make it easier
for us to track such issues. For the time being we only apply it to
2 testcases, but it would be nice to track some of the other tests that
were skipped in a similar fashion and apply it to them as well (instead
of skipIfLinux). Different patch though.