[Morello] Don't adjust sections for representability during relocatable link
Most of this operation relies on section VAs, which are completely
meaningless during a relocatable link (they will be assigned, but later
all reset back to 0), and the entire layout will be reshuffled anyway
when linked into the final binary. Moreover, if the last section in the
"PCC segment" ends up being an SHF_MERGE section then we end up messing
with its size and risk rendering its size not a multiple of its element
size. This has been seen to give errors like:

  ld.lld: error: pw.lo:(.rodata.cst32): SHF_MERGE section size (40) must be a multiple of sh_entsize (32)

when building CheriBSD, where pw.lo is the output of ld -r'ing all the
object files for the pw binary together and in this case is being linked
into cheribsdbox to produce a crunched binary (like busybox in the Linux
world). For whatever reason, this only started showing up after merging
mid-summer CHERI LLVM into Morello LLVM, presumably due to changes in
LLD's layout algorithms that started to expose this latent bug.