[Morello] Use integer rather than capability branch to lower jump tables
On Rainier, PCC bounds are not predicted, and only readable by ADR(P)
once the branch has been resolved. This means that, if a jump table is
in a tight loop, the backend fills up with ADR->BR->ADR->BR->... chains
that result in large numbers of backend stalls. In particular, the state
machine benchmark in CoreMark hits this, resulting in a 15% overhead
(cycles) and around 3x the backend stalls that account for almost all
this overhead. Using indirect integer branches, whilst sad that it has
to be done to achieve competitive performance, should be just as safe,
and reduces CoreMark's overhead to around 1% more cycles.