[Morello][lldb] Fix the number of reg infos
When sending the register set, lldb-server doesn't manage to fetch
the registers above the first capability register. This is because when
iterating through the registers it's using
RegisterContext::GetUserRegisterCount as the upper limit, but when
fetching the RegisterInfo for each index it's checking against
RegisterInfoPOSIX_arm64::GetRegisterCount. This is set in the
constructor as fpu_fpcr + 1, but that doesn't include the capability
registers, which begin right after that.

This ugly patch fixes the constructor of RegisterInfoPOSIX_arm64 to
agree with GetUserRegisterCount, so we recognize higher indices as
valid. A more principled patch should clean these interfaces up, but
that's something we should do upstream, not here.