SelectionDAG::FoldConstantArithmetic: Handle different vector VTs
When folding vector operations, the current code assumed that both LHS and
RHS vectors had matching value types. However, for a vector PTRADD node,
the LHS is a iFATPTR128 and the RHS is i64, so we can't just create an
undef node of iFATPTR128 type for the RHS. This commit fixes the problem
by using the VT of both nodes instead of assuming they match the result VT.

It probably also makes sense to commit this to CHERI LLVM (without the test
case) even though the case can't currently be triggered as vector PTRADD
is not supported for RISC-V yet:

```
SplitVectorOperand Op #1: t13: iFATPTR128 = PTRADD t4, t12

LLVM ERROR: Do not know how to split this operator's operand!
```