[Morello][lldb] Read the ABI from ELF
Distinguish between purecap and hybrid based on the existence of the
EF_AARCH64_CHERI_PURECAP flag in the binary. This information is passed
into the ArchSpec and used when determining the size of a pointer. Note
that even with this patch we still don't have a fully consistent
handling of pointers vs addresses in lldb, so certain things might still
be broken. However, we can now enable one of the end-to-end tests for
hybrid mode. It used to pass only for purecap because we were making an
assumption about certain pointers being capabilities.