[Morello][lldb] Run end-to-end tests with 2 ABIs
The end-to-end lldb tests used to be run only under the purecap ABI.
This patch makes it possible to build and run a test both as purecap and
as hybrid. It also enables most tests to be run for both ABIs. In order
to do so, it sometimes adds `__capability` qualifiers to pointers (e.g.
in order to trigger a capability fault even under the hybrid ABI).

See the updates to the README for the low level details.