[clang] Remove unnecessary __sync_fetch_and_(u)min/max overload
Upstream Clang only supports this extension on ints, as does CHERI LLVM.
For some unknown reason Morello LLVM has overloaded this since the first
commit, but this does not seem necessary. Real world code won't be using
it on anything other than ints as a result, and so it's already not
usable on pointers, thus not a concern for CHERI that there's no
capability version. It's also unlikely anyone's really using it since
it's a Clang extension for the deprecated __sync builtins. Upstreaming
the overloading might make sense if the __sync builtins weren't
deprecated, but nobody should be writing new code with them when better
alternatives exist (GNU __atomic and Clang __c11_atomic intrinsics or
standard C(++)11 atomics).