[Cheri] Use a sub with cheri.cap.address.get operands instead of a cheri.cap.diff. This better matches what non-cheri code does, and allows the optimizers to further act on the address.get and and sub.
Limit this to the case where we're not using capability
offsets for intcap to not impact the garbage collection
use case.

Stop instcombine from creating a cheri.cap.diff when
seeing a subtract of cheri.cap.address.get. To compensate
for the existing cheri.cap.diff optimization not being
active anymore (folding to a constant where possible),
add the corresponding optimization for a sub of
cheri.cap.address.get instructions.