[Morello][SCEV] Fix expansion for add SCEV expressions.
Check OpsAndLoops instead of Loops to find out the number
of operands of the expression. This was causing an issue
where we weren't doing the usual base/offset expansion.

Additionally fix the logic in expandAddToGEP for capabilities.
Just emit the ugly GEP to avoid the case where we would need
multiple GEPs.

Fixes https://git.morello-project.org/morello/llvm-project/-/issues/65