[Morello][lldb] Fix reading uninitialised register in test
Nothing initialises c11 so whilst its address and bounds get set the
rest of its metadata is completely unspecified, with the current CHECK
lines just checking for whatever the current code generation for the
current Android start-up code sequence and main function prologue
produces. A future code generation change will perturb this enough to
change the non-bounds metadata seen in c11 here, so fix this test to
fully initialise the capability.

As part of this, tighten up the CHECK lines, and note that the output
being tested for actually highlights a bug in Morello LLDB's bounds
decoding (using the full address for bounds calculations rather than
sign-extending the low 56 bits).