[Morello] Start to deprecate the usage of c64 in the arch string
c64 is the mode, and should not be part of the arch string. This
starts the migration for removing c64 entirely from the arch string.

a64c in the arch string enables capabilities support and is orthogonal
to the CPU mode.

For assembly, we're adding two new directives to change the mode to
a64 or c64 respectively:
  .code a64
  .code c64
These should be used to change the mode instead of .arch or
.arch_extension.

Usage of c64 in the march option that is passed to clang will now give
a warning.

In the future any usage of c64 in the arch string will produce a warning,
(and eventually an error) both as part of -march or in .arch or
.arch_extension directives. Any mode changes with .arch or .arch_extension
should also warn (and eventually this behaviour should be deprecated).
However this is not yet the default, as we need the new mode changing
directives to be widely supported.

For command line options, the c64 mode can already be inferred from the ABI
and can be usually ommitted from the march string. We now warn when
we find c64 in the march string.